+++ nens_auth_client/checks.py
@@ -62,3 +62,31 @@ def check_access_token_middleware_order(app_configs=None, **kwargs):
             )
         ]
     return []
+
+
+@register()
+def check_error_message_formatting(app_configs=None, **kwargs):
+    """Check ERROR_INVITATION_WRONG_USER and _EMAIL for correct formatting.
+
+    'Error message' is accepted, but 'Error message {invalid_placeholder}' not.
+    """
+    errors = []
+    fmt = settings.NENS_AUTH_ERROR_INVITATION_WRONG_USER
+    try:
+        fmt.format(expected_user="a", actual_user="b")
+    except KeyError:
+        errors.append(Error(
+            "Ensure that NENS_AUTH_ERROR_INVITATION_WRONG_USER only contains "
+            "'expected_user' and 'actual_user' placeholders."
+        ))
+
+    fmt = settings.NENS_AUTH_ERROR_INVITATION_WRONG_EMAIL
+    try:
+        fmt.format(expected_email="a", actual_email="b")
+    except KeyError:
+        errors.append(Error(
+            "Ensure that NENS_AUTH_ERROR_INVITATION_WRONG_EMAIL only contains "
+            "'expected_email' and 'actual_email' placeholders."
+        ))
+
+    return errors